src/main/java/com/aydin/account/service/CustomerService.java
package com.aydin.account.service;

import com.aydin.account.dto.CustomerDto;
import com.aydin.account.dto.converter.CustomerDtoConverter;
import com.aydin.account.exception.CustomerNotFoundException;
import com.aydin.account.model.Customer;
import com.aydin.account.repository.CustomerRepository;
import org.springframework.stereotype.Service;


@Service
public class CustomerService {

    private final CustomerRepository customerRepository;
    private final CustomerDtoConverter customerDtoConverter;
    public CustomerService(CustomerRepository customerRepository, CustomerDtoConverter customerDtoConverter) {
        this.customerRepository = customerRepository;
        this.customerDtoConverter = customerDtoConverter;
    }

    protected Customer findCustomerById(String id) {
        return customerRepository.findById(id)
                .orElseThrow(() -> new CustomerNotFoundException("Customer mot found by id:  " + id));
    }

    public CustomerDto getCustomerById(String customerId) {
        return customerDtoConverter.convertToCustomerDto(findCustomerById(customerId));


    }
}
